backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
asyncpg>=0.30.0
PyJWT>=2.9.0
httpx>=0.28.0
playwright>=1.49.0
beautifulsoup4>=4.12.3
jsonschema>=4.23.0
prometheus-client>=0.21.0
pydantic>=2.11.0
pydantic-settings>=2.8.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
